fix: reorganize code a bit
