+++ api/v1alpha1/enterprise_types.go
@@ -5,7 +5,7 @@ package v1alpha1
 import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	"github.com/mercedes-benz/garm-operator/pkg/util/conditions"
+	"github.com/mercedes-benz/garm-operator/pkg/conditions"
 )
 
 // EnterpriseSpec defines the desired state of Enterprise
